platform/backend/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
decouple==0.0.7
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
Markdown==3.3.4
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
